requirements.txt
kafka-python
requests
pika
hurry.filesize
tqdm